+++ R/package.R
@@ -66,6 +66,7 @@
 #' @importFrom stringr str_split str_trim
 #' @importFrom testthat context expect_false expect_true test_dir test_that
 #' @importFrom tibble as_tibble tibble
+#' @importFrom tidyselect vars_select
 #' @importFrom utils capture.output compareVersion installed.packages
 #'   packageVersion read.csv sessionInfo type.convert unzip write.table
 #' @importFrom visNetwork toVisNetworkData visEvents visHierarchicalLayout